+++ lib/exceptions/socks_client_exception.dart
@@ -1,3 +1,0 @@
-class SocksClientException implements Exception {
-
-}