+++ lib/socks.dart
@@ -1,3 +1,3 @@
+export 'enums.dart';
 export 'socks_client.dart';
-export 'socks_enums.dart';
 export 'socks_server.dart';