+++ lib/src/client/socks_client.dart
@@ -5,10 +5,12 @@ import 'dart:typed_data';
 import 'package:async/async.dart';
 import 'package:meta/meta.dart';
 
-import '../../enums/authentication_method.dart';
-import '../../enums/command_reply_code.dart';
-import '../../enums/socks_connection_type.dart';
+import '../../exceptions.dart';
 import '../address_type.dart';
+import '../enums/authentication_method.dart';
+import '../enums/command_reply_code.dart';
+import '../enums/socks_connection_type.dart';
+import '../exceptions/client/connection_command_failed_exception.dart';
 import '../mixin/byte_reader.dart';
 import '../mixin/socket_mixin_.dart';
 import '../mixin/stream_mixin.dart';
@@ -33,7 +35,7 @@ class SocksSocket with StreamMixin<Uint8List>, SocketMixin, ByteReader {
 
   final SocksConnectionType type;
 
-  /// Can be overriden/set to be custom domain lookup function.
+  /// Can be overridden/set to be custom domain lookup function.
   LookupFunction lookup = InternetAddress.lookup;
 
   @override
@@ -57,27 +59,35 @@ class SocksSocket with StreamMixin<Uint8List>, SocketMixin, ByteReader {
     int port,
     SocksConnectionType type,
   ) async {
-    if(proxies.isEmpty)
+    if(proxies.isEmpty) {
       throw ArgumentError.value(proxies, 'proxies', 'empty');
+    }
 
     final socket = await Socket.connect(proxies.first.host, proxies.first.port);
   
     final client = SocksSocket.protected(socket, type);
-    await client._handshake(proxies.first);
 
-    for(var i = 1; i < proxies.length; i++) {
-      await client._handleCommand(proxies[i].host, proxies[i].port, SocksConnectionType.connect);
-      final response = await client._handleCommandResponse(SocksConnectionType.connect);
-      if(response.address != InternetAddress('0.0.0.0') || response.port != 0)
-        throw UnimplementedError('Connect associated proxy not yet implemented.');
-      await client._handshake(proxies[i]);
+    try {
+      await client._handshake(proxies.first);
+
+      for(var i = 1; i < proxies.length; i++) {
+        await client._handleCommand(proxies[i].host, proxies[i].port, SocksConnectionType.connect);
+        final response = await client._handleCommandResponse(SocksConnectionType.connect);
+        if(response.address != InternetAddress('0.0.0.0') || response.port != 0) {
+          throw UnimplementedError('Connect associated proxy not yet implemented.');
+        }
+        await client._handshake(proxies[i]);
+      }
+
+      await client._handleCommand(address, port, type);
+
+      final response = await client._handleCommandResponse(type);
+      
+      return SocksClientInitializeResult(client, response);
+    } on ByteReaderException catch (error, stackTrace) {
+      socket.close().ignore();
+      throw SocksClientConnectionClosedException((error: error, stackTrace: stackTrace));
     }
-
-    await client._handleCommand(address, port, type);
-
-    final response = await client._handleCommandResponse(type);
-
-    return SocksClientInitializeResult(client, response);
   }
 
   // Apply tls-over-http
@@ -154,7 +164,7 @@ class SocksSocket with StreamMixin<Uint8List>, SocketMixin, ByteReader {
     // Checking authentication version.
     if (await readUint8() != 0x01) {
       close().ignore();
-      throw Exception('Unsupported userpass authentication version.');
+      throw Exception('Unsupported user/pass authentication version.');
     }
     // Checking authentication response, 0x00 - succeed, other - failed.
     if (await readUint8() != 0x00) {
@@ -197,9 +207,7 @@ class SocksSocket with StreamMixin<Uint8List>, SocketMixin, ByteReader {
  
     if (commandResponse != CommandReplyCode.succeed) {
       close().ignore();
-      throw Exception(
-        'Command handling failed. With error: ${commandResponse.name}',
-      );
+      throw SocksClientConnectionCommandFailedException(commandResponse);
     }
 
     // Read reserved byte.