+++ lib/src/client/socks_tcp_client.dart
@@ -1,12 +1,11 @@
-
-
 import 'dart:io';
 
-import '../../enums/socks_connection_type.dart';
+import '../enums/socks_connection_type.dart';
 import '../shared/proxy_settings.dart';
-import 'socket_connection_task.dart';
 import 'socks_client.dart';
 
+
+/// [Socket] wrapper for socks TCP connection.
 class SocksTCPClient extends SocksSocket {
   SocksTCPClient._internal(Socket socket) : super.protected(socket, SocksConnectionType.connect);
 
@@ -42,16 +41,20 @@ class SocksTCPClient extends SocksSocket {
         );
         
         // Secure connection after establishing Socks connection
-        if(uri.scheme == 'https')
-          return SocketConnectionTask((await client).secure(uri.host, 
+        if(uri.scheme == 'https') {
+          final Future<SecureSocket> secureClient;
+          return ConnectionTask.fromSocket(secureClient = (await client).secure(
+            uri.host, 
             context: context,
             onBadCertificate: onBadCertificate,
             keyLog: keyLog,
             supportedProtocols: supportedProtocols,
-          ),);
+          ), () async => (await secureClient).close().ignore(),);
+        }
 
         // SocketConnectionTask implements ConnectionTask<Socket>
-        return SocketConnectionTask(client);
+        return ConnectionTask.fromSocket(client, 
+          () async => (await client).close().ignore(),);
       };
   }