+++ lib/src/exceptions/byte_reader_exception.dart
@@ -1,0 +1,8 @@
+import '../mixin/byte_reader.dart';
+
+/// Exception thrown by [ByteReader] mixin
+class ByteReaderException implements Exception {
+  const ByteReaderException(this.message);
+
+  final String message;
+}